Add select input with functionality
Adds a select input with three options: The first one is disabled and
asks the user if they have read the book. The second one is 'Read' with
the value 'true' and the last one is 'Not read' with the value 'false'.

Styling, validation and error messages are added like the other inputs.
However, an extra class had to be added to the other inputs' labels to
distinguish them from the select label. Otherwise all labels were
assumed to contain an input element, which could lead to TypeErrors.