Implement functions for adding books
Adds the example book to the library array.

Checks for an identical (case insensitive) book title before adding a
book. Adds an htmlMarkup propety to the book with template literals
for use with the insertAdjacentHTML function, and a toggleStatus
function property for use with a toggle button.

Also adds functions for displaying one book, and displaying the entire
library.