Add remove button, closes #2
A div with the class remove is added to both the list headers and each
book to make sure elements are lined up correctly. A button is added to
div.remove in the book with a 'click' eventListener.

Clicking the button first asks the user for a confirmation. If accepted,
the function removeBook(book) is called. This function removes a book
from the library by using a for-loop, and removes it from the DOM by
using querySelectorAll and a forEach-loop. In both cases, the title is
used to find the correct element and the loop only runs until the
element is found.