Add title attribute, simplify removing books
With the title attribute, removing a book with a specific title becomes
much easier.

Before this, querySelectorAll('.book') returned a NodeList with all
books. Another querySelector had to be used to find .title and finally
textContent could be compared to the title of the book that should be
removed.

Now, only one querySelector has to be used, which looks for a book with
title="${book.title}". This book can then be immediately removed.

This also fixes a problem where the textContent is changed by a third
party program such as 'Toucan'. In that case, searching for the book
would not work.