Implement first version of sorting, part of #3
Uses EventListeners for the list headers that call a sorting function.
Depending on which header is clicked, the list will be sorted either
based on author (then title), status (then title), pages (then title) or
simply title. After sorting the list, the DOM is updated.

However, the code is unncessesarily complicated. As all the functions
eventually sort based on title, there may be an easier way to implement
the sorting.