Fix status header alignment when not sorting
The arrow took up space with visibility: hidden, which made the status
alignment incorrect when not sorting based on status.

Instead display: none and display: block are used so the arrows are only
showed when sorting based on status.

The .status selector is also changed to div.status to increase the
specificity and allow justify-content: flex-end to apply.